feat: decouple account selector from qr code connector
In order to enable account selector being reusable, we have to decouple the account selector from QR hardware wallet

the changes including decouple account selector & account detail components

in addtional
we also add:
An hook useBlockExplorer for generate the block explorer redirect method
An hook useAccountsBalance for generate the account balance mapping
