+++ app/components/UI/HardwareWallet/AccountSelector/hooks.tsx
@@ -1,0 +1,42 @@
+import { useState, useMemo, useEffect } from 'react';
+import Engine from '../../../../core/Engine';
+
+export interface IAccount {
+  address: string;
+}
+
+export interface AccountBalance {
+  balance: string;
+}
+
+export interface AccountBalances {
+  [p: string]: AccountBalance;
+}
+
+export const useAccountsBalance = (accounts: IAccount[]) => {
+  const [trackedAccounts, setTrackedAccounts] = useState<AccountBalances>({});
+  const AccountTrackerController = useMemo(
+    () => (Engine.context as any).AccountTrackerController,
+    [],
+  );
+
+  useEffect(() => {
+    const unTrackedAccounts: string[] = [];
+    accounts.forEach((account) => {
+      if (!trackedAccounts[account.address]) {
+        unTrackedAccounts.push(account.address);
+      }
+    });
+    if (unTrackedAccounts.length > 0) {
+      AccountTrackerController.syncBalanceWithAddresses(unTrackedAccounts).then(
+        (_trackedAccounts: AccountBalances) => {
+          setTrackedAccounts(
+            Object.assign({}, trackedAccounts, _trackedAccounts),
+          );
+        },
+      );
+    }
+  }, [AccountTrackerController, accounts, trackedAccounts]);
+
+  return trackedAccounts;
+};
