+++ app/components/UI/HardwareWallet/AccountSelector/types.tsx
@@ -2,6 +2,6 @@ export interface IAccount {
   address: string;
   balance: string;
   index: number;
-  checked: boolean;
-  exist: boolean;
+  checked?: boolean;
+  exist?: boolean;
 }
