+++ locales/languages/el.json
@@ -581,11 +581,13 @@
     "purpose_connect": "συνδεθείτε",
     "purpose_sign": "επιβεβαιώσετε τη συναλλαγή",
     "select_accounts": "Επιλέξτε Λογαριασμό",
+    "please_wait": "Παρακαλούμε περιμένετε"
+  },
+  "account_selector": {
     "prev": "ΠΡΟΗΓ",
     "next": "ΕΠΌΜΕΝΟ",
     "unlock": "Ξεκλειδωμα",
-    "forget": "Διαγραφή αυτής της συσκευής",
-    "please_wait": "Παρακαλούμε περιμένετε"
+    "forget": "Διαγραφή αυτής της συσκευής"
   },
   "app_settings": {
     "title": "Ρυθμίσεις",
