+++ locales/languages/es.json
@@ -581,11 +581,13 @@
     "purpose_connect": "conectar",
     "purpose_sign": "confirmar la transacción",
     "select_accounts": "Seleccionar una cuenta",
+    "please_wait": "Espere, por favor"
+  },
+  "account_selector": {
     "prev": "PREV",
     "next": "SIGUIENTE",
     "unlock": "Desbloquear",
-    "forget": "Olvidar este dispositivo",
-    "please_wait": "Espere, por favor"
+    "forget": "Olvidar este dispositivo"
   },
   "app_settings": {
     "title": "Configuración",
