+++ locales/languages/id.json
@@ -581,11 +581,13 @@
     "purpose_connect": "hubungkan",
     "purpose_sign": "konfirmasikan transaksi",
     "select_accounts": "Pilih Akun",
+    "please_wait": "Harap tunggu"
+  },
+  "account_selector": {
     "prev": "SEBELUMNYA",
     "next": "SELANJUTNYA",
     "unlock": "Buka",
-    "forget": "Lupakan perangkat ini",
-    "please_wait": "Harap tunggu"
+    "forget": "Lupakan perangkat ini"
   },
   "app_settings": {
     "title": "Pengaturan",
