+++ locales/languages/ko.json
@@ -581,11 +581,13 @@
     "purpose_connect": "연결",
     "purpose_sign": "트랜잭션 컨펌",
     "select_accounts": "계정 선택",
+    "please_wait": "잠시 기다리세요"
+  },
+  "account_selector": {
     "prev": "이전",
     "next": "다음",
     "unlock": "잠금 해제",
-    "forget": "이 장치 잊기",
-    "please_wait": "잠시 기다리세요"
+    "forget": "이 장치 잊기"
   },
   "app_settings": {
     "title": "설정",
