+++ locales/languages/pt.json
@@ -581,11 +581,13 @@
     "purpose_connect": "conectar",
     "purpose_sign": "confirmar a transação",
     "select_accounts": "Selecione uma conta",
+    "please_wait": "Aguarde"
+  },
+  "account_selector": {
     "prev": "VOLTAR",
     "next": "AVANÇAR",
     "unlock": "Desbloquear",
-    "forget": "Esquecer este dispositivo",
-    "please_wait": "Aguarde"
+    "forget": "Esquecer este dispositivo"
   },
   "app_settings": {
     "title": "Configurações",
