+++ locales/languages/ru.json
@@ -581,11 +581,13 @@
     "purpose_connect": "подключиться",
     "purpose_sign": "подтвердить транзакцию",
     "select_accounts": "Выберите счет",
+    "please_wait": "Пожалуйста, подождите..."
+  },
+  "account_selector": {
     "prev": "ПРЕДЫДУЩИЙ",
     "next": "ДАЛЕЕ",
     "unlock": "Разблокировать",
-    "forget": "Забыть это устройство",
-    "please_wait": "Пожалуйста, подождите..."
+    "forget": "Забыть это устройство"
   },
   "app_settings": {
     "title": "Настройки",
