+++ locales/languages/vi.json
@@ -581,11 +581,13 @@
     "purpose_connect": "kết nối",
     "purpose_sign": "xác nhận giao dịch",
     "select_accounts": "Chọn một tài khoản",
+    "please_wait": "Vui lòng chờ"
+  },
+  "account_selector": {
     "prev": "TRƯỚC",
     "next": "TIẾP",
     "unlock": "Mở khóa",
-    "forget": "Quên thiết bị này",
-    "please_wait": "Vui lòng chờ"
+    "forget": "Quên thiết bị này"
   },
   "app_settings": {
     "title": "Cài đặt",
