+++ locales/languages/zh.json
@@ -587,6 +587,12 @@
     "forget": "忽略此设备",
     "please_wait": "请稍候"
   },
+  "account_selector": {
+    "prev": "上一步",
+    "next": "下一步",
+    "unlock": "登录",
+    "forget": "忽略此设备"
+  },
   "app_settings": {
     "title": "设置",
     "current_conversion": "基础货币",
